requirements.txt
py-cord == 2.2.2
discord.py==2.4.0